Update the extension to work with the updated daisuki page
Since daisuki now loads the swf from an iframe, the extension has been modified to update the
page on the iframe being loaded.

Additionally for some reason agcrypto and bgnwrapper isn't always immediately loaded and available
when the iframe is loaded.

This patch just directly loads the scripts using jquery before attempting to update the page.